third_party/WebKit/LayoutTests/media/video-currentTime-before-have-metadata-media-fragment-uri-expected.txt

EXPECTED (video.currentTime == '0') OK
EVENT(loadedmetadata)
EXPECTED (video.currentTime == '1') OK
EVENT(seeked)
END OF TEST
